Moved models out to their own file
Also, a little cleaning up of docs
